Reorganise location and naming of files
