+++ projects/archiver/main.ts
@@ -1,9 +1,9 @@
 require('dotenv').config()
 export { handler as invoke } from './src/invoke'
-export { handler as issue } from './src/tasks/issueTask'
-export { handler as front } from './src/tasks/frontTask'
-export { handler as image } from './src/tasks/imageTask'
-export { handler as upload } from './src/tasks/issueUploadTask'
-export { handler as zip } from './src/tasks/zipTask'
-export { handler as indexer } from './src/tasks/generateIndexTask'
-export { handler as event } from './src/tasks/notificationTask'
+export { handler as issue } from './src/tasks/issue-task'
+export { handler as front } from './src/tasks/front-task'
+export { handler as image } from './src/tasks/image-task'
+export { handler as upload } from './src/tasks/issue-upload-task'
+export { handler as zip } from './src/tasks/zip-task'
+export { handler as indexer } from './src/tasks/generate-index-task'
+export { handler as event } from './src/tasks/notification-task'
