+++ projects/archiver/src/tasks/front-task.ts
@@ -2,13 +2,13 @@ import { Handler } from 'aws-lambda'
 import { unnest } from 'ramda'
 import { attempt, hasFailed } from '../../../backend/utils/try'
 import { frontPath, Image } from '../../common'
-import { getImagesFromFront } from '../../media'
-import { getFront } from '../downloader'
-import { IssueTaskOutput } from './issueTask'
-import { Bucket } from '../s3'
-import { upload, ONE_WEEK } from '../upload'
-import { logInput, logOutput } from '../log-utils'
-import { handleAndNotifyOnError } from '../notifications/pub-status-notifier'
+import { getImagesFromFront } from './image/media'
+import { getFront } from '../utils/backend-client'
+import { IssueTaskOutput } from './issue-task'
+import { Bucket } from '../utils/s3'
+import { upload, ONE_WEEK } from '../utils/s3'
+import { logInput, logOutput } from '../utils/log'
+import { handleAndNotifyOnError } from './notifications/pub-status-notifier'
 
 type FrontTaskInput = IssueTaskOutput
 export interface FrontTaskOutput extends IssueTaskOutput {
