+++ projects/archiver/src/tasks/indexer/get-issues.ts
@@ -1,5 +1,5 @@
-import { IssueIdentifier } from '../../../common/src'
-import { Bucket, listNestedPrefixes } from '../s3'
+import { IssueIdentifier } from '../../../../common/src'
+import { Bucket, listNestedPrefixes } from '../../utils/s3'
 
 /* Crawl S3 for a list of all of the issues that are available */
 export const getIssues = async (): Promise<IssueIdentifier[]> => {
