+++ projects/archiver/src/tasks/indexer/get-published-version.ts
@@ -1,8 +1,8 @@
 import {
     IssuePublicationIdentifier,
     IssueIdentifier,
-} from '../../../common/src'
-import { getStatuses, isPublished } from '../status'
+} from '../../../../common/src'
+import { getStatuses, isPublished } from '../../status-store/status'
 
 /* Given an edition name and date this will return the current publication instance ID
  * or undefined is there is no valid published instance. This is based on the status
