+++ projects/archiver/src/utils/path-builder.ts
@@ -1,4 +1,4 @@
-import { IssuePublicationIdentifier, IssueIdentifier } from '../common'
+import { IssuePublicationIdentifier, IssueIdentifier } from '../../common'
 
 export const getPublishedId = (issuePublication: IssuePublicationIdentifier) =>
     `${issuePublication.edition}/${issuePublication.issueDate}/${issuePublication.version}`
